fix use 64bit for time

+++ src/binance/proxy.h
@@ -43,7 +43,7 @@ public:
 
 
 private:
-	std::intptr_t time_diff = 0;
+	std::int64_t time_diff = 0;
 	mutable std::uint64_t time_sync = 0;
 	void buildParams(const json::Value& params, std::ostream& data);
 };
